Fix harmless warning.
util.c: In function p_info:
util.c:157: warning: format %08ld expects type long int, but argument 4 has type __darwin_suseconds_t

+++ util.c
@@ -154,7 +154,7 @@ void p_info(const char *fmt, ...) {
 	if (DEBUG > 0) {
 		struct timeval tv;
 		gettimeofday(&tv, NULL);
-		fprintf(stdout, "%08ld.%08ld ", tv.tv_sec, tv.tv_usec);
+		fprintf(stdout, "%08ld.%08ld ", (long)tv.tv_sec, (long)tv.tv_usec);
 
 		char date[64];
 		struct tm tm;
